# api/openapi.yaml
openapi: 3.0.3
info:
  title: Point of Service (POS) API
  description: REST API to be used to create tableside food ordering applications.
  contact:
    name: Course Discord server
    url: https://discord.gg/dPsnz5u9
    email: mark.wentowski@docsgeek.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - url: http://localhost:80/
    description: Local setup
  - url: '{gitpod-server-url}/'
    description: Gitpod setup
    variables:
      gitpod-server-url:
        default: 'https://8080-<workspace-id>.gitpod.io'
        description: 'Dynamic Gitpod URL. Replace <workspace-id> with your Gitpod workspace ID if needed.'
...
paths:
...
  /ingredients:
    get:
      tags:
        - Ingredients
      summary: Get a list of ingredients
      description:
        "This operation retrieves a list of ingredients based on the specified\
        \ criteria. You can sort the results, include specific properties, select\
        \ ingredients that match certain criteria, and paginate through the results."
      operationId: getIngredientList
      parameters:
        - name: sort
          in: query
          description: |
            Specifies the field by which the results should be sorted. This parameter allows you to define the sorting order based on a particular field in the dataset. If you include a `sort` value, you must include a `order` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name
        - name: order
          in: query
          description: |
            Determines the order in which the results should be sorted. This parameter specifies whether the sorting should be in ascending or descending order. Valid values are `asc` for ascending order and `desc` for descending order. If you include an `order` value, you must include a `sort` value.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: asc
        - name: fields
          in: query
          description: |
            Use this parameter to specify which resource properties should be included in the response. Provide a list of schema property names, separated by the `|` delimiter. This allows for precise control over the data returned.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: name|in_stock_qty
        - name: filter
          in: query
          description: |
            Selection filters to display resources matching the provided criteria. Use the format `fieldName.operator~value`.

            **Basic Equality (eq)**
            - For exact matches, use `fieldName.eq~value`.

            **Comparison Operators:**
            - **Less Than (lt)** Use `fieldName.lt~value`.
            - **Less Than or Equal To (lte)** Use `fieldName.lte~value`.
            - **Greater Than (gt):** Use `fieldName.gt~value`.
            - **Greater Than or Equal To (gte):** Use `fieldName.gte~value`.

            Separate filters using the `|` delimter. Dates should be in `YYYY-MM-DD` format.
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: created_at.gt~2024-05-07|in_stock_qty.lt~5
        - name: limit
          in: query
          description: |
            Pagination filter for the maximum number of results to return in the response.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            type: integer
          example: 10
        - name: offset
          in: query
          description: |
            Offest-based pagination filter for the number of results to skip before displaying resources.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Ingredients
    post:
      tags:
        - Ingredients
      summary: Create an ingredient
      description:
        "This operation creates a new ingredient with the provided information.\
        \ Set an initial quantity of the ingredient using the `in_stock_qty` field.\n\
        \    \nThe `id` returned in the response can be used to associate the ingredient\
        \ with a dish. The `in_stock_qty` can be used to track if the ingredient has\
        \ a positive quantity.\n"
      operationId: postIngredient
      requestBody:
        $ref: '#/components/requestBodies/CreateIngredientRequest'
      responses:
        '201':
          description: |
            Resource created successfully. The response body contains the newly created resource.
          headers:
            Location:
              description: Points to the resource endpoint for the newly created resource
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Ingredients
  /ingredients/{id}:
    get:
      tags:
        - Ingredients
      summary: Get an ingredient's details
      description:
        This operation retrieves information about a specific ingredient
        identified by its unique ingredient ID.
      operationId: getIngredient
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 614d706b67c122001f35cc47
        - name: fields
          in: query
          description: |
            Projection filters for including only certain resource properties in the response. The value for the parameter should be a comma-separated list of properties from the schema.
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: in_stock_qty|name
      responses:
        '200':
          description:
            Successful operation. The response body contains the requested
            data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      x-swagger-router-controller: Ingredients
    put:
      tags:
        - Ingredients
      summary: Update an ingredient
      description:
        "This operation updates an ingredient using the provided information.\
        \ Using this operation, API consumers can update the `in_stock_qty` of ingredients\
        \ when orders are placed."
      operationId: putIngredient
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 617d2e09ec42e0001f41bf84
      requestBody:
        $ref: '#/components/requestBodies/UpdateIngredientRequest'
      responses:
        '200':
          description:
            Successful operation. The response body contains the updated
            resource.
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description:
            The request could not be completed due to a conflict with the
            current state of the resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Ingredients
    delete:
      tags:
        - Ingredients
      summary: Delete an ingredient
      description:
        This operation deletes an existing order identified by its unique
        order ID.
      operationId: deleteIngredient
      parameters:
        - name: id
          in: path
          description: The unique identifier for the resource.
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 61c8c34bb02368001f8f0812
      responses:
        '204':
          description:
            "No content. The operation was successful, and there is no\
            \ content to return."
        '400':
          description: Bad request. The request was malformed or invalid.
        '401':
          description:
            "Unauthorized. Authentication is required, or the provided\
            \ credentials are invalid."
        '403':
          description:
            "Forbidden. The server understood the request, but it refuses\
            \ to authorize it. Authentication will not help."
        '404':
          description: Not found. The requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description:
            Internal server error. An unexpected error occurred on the
            server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
        - AccessToken: []
      x-swagger-router-controller: Ingredients
...
components:
  schemas:
...
    CreatedAt:
      type: string
      description: Timestamp when the resource was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    Id:
      type: string
      description: The unique identifier for the resource.
      readOnly: true
      example: 674d0bf5c28b69001f8e03a1
...
    Ingredient:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        created_at:
          $ref: '#/components/schemas/CreatedAt'
        updated_at:
          $ref: '#/components/schemas/UpdatedAt'
        name:
          type: string
          description: A human-friendly name for the ingredient.
          example: Carrot
        in_stock_qty:
          $ref: '#/components/schemas/InStockQty'
        price:
          $ref: '#/components/schemas/Price'
      example:
        updated_at: 2024-04-22T10:00:00Z
        price: 999
        name: Carrot
        created_at: 2024-04-22T10:00:00Z
        id: 674d0bf5c28b69001f8e03a1
        in_stock_qty: 3
    InStockQty:
      type: integer
      description:
        The quantity of the ingredient in the restaurant's inventory. The
        quantity of the ingredient should be managed by the API consumer.
      example: 3
...
    Price:
      type: number
      description:
        The cost of the dish in cents. For example `999` equals `9.99`
        and `12999` equals `129.99`.
      format: float
      example: 999
...
    Problem:
      required:
        - detail
        - title
      type: object
      properties:
        title:
          type: string
          description: Human-readable error title.
          example: Human-readable error title.
        detail:
          type: string
          description: Human-readable error details.
          example: Human-readable error details.
...
    UpdatedAt:
      type: string
      description: Timestamp when the order was created in ISO 8601 format.
      format: date-time
      readOnly: true
      example: 2024-04-22T10:00:00Z
...
    inline_response_200_4:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        total_results:
          type: integer
          example: 1
      example:
        results:
          - updated_at: 2024-04-22T10:00:00Z
            price: 999
            name: Carrot
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            in_stock_qty: 3
          - updated_at: 2024-04-22T10:00:00Z
            price: 999
            name: Carrot
            created_at: 2024-04-22T10:00:00Z
            id: 674d0bf5c28b69001f8e03a1
            in_stock_qty: 3
        total_results: 1
...
  requestBodies:
...
    CreateIngredientRequest:
      description: Request body containing information used to create the resource.
      content:
        application/json:
          schema:
            required:
              - in_stock_qty
              - name
              - price
            type: object
            properties:
              name:
                type: string
                description: A human-friendly name for the ingredient.
                example: Carrot
              in_stock_qty:
                $ref: '#/components/schemas/InStockQty'
              price:
                $ref: '#/components/schemas/Price'
      required: true
    UpdateIngredientRequest:
      description: Request body containing information used to update the resource.
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                description: A human-friendly name for the ingredient.
                example: Carrot
              in_stock_qty:
                $ref: '#/components/schemas/InStockQty'
              price:
                $ref: '#/components/schemas/Price'
      required: true
...
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      description: |
        Authentication using client credentials. Use the `client_id` as the username and `client_secret` as the password.

        - **client_id**: Public identifier for the client application. Example: `admin_web_app`.
        - **client_secret**: Confidential string for client authentication. Example: `a5d7f23e-8b64-4b4c-9b11-21c5cfdf25f1`.
      scheme: basic